# api-docs.yaml
openapi: 3.0.0
info:
  title: ClickUp Webhook Manager
  version: 1.0.0
  description: API for managing ClickUp webhooks
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local Development Server

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token de API do ClickUp

  schemas:
    WebhookCreate:
      type: object
      required:
        - endpoint
      properties:
        endpoint:
          type: string
          example: "https://localhost:8000/team/{team_id}/webhook"
        description:
          type: string
          example: "Webhook for task updates"

    Webhook:
      type: object
      properties:
        id:
          type: string
          example: "webhook_123"
        endpoint:
          type: string
          example: "https://localhost:8000/team/{team_id}/webhook"
        client_id:
          type: string
          example: "client_123"
        workspace_id:
          type: string
          example: "workspace_123"
        user_id:
          type: string
          example: "user_123"
        events:
          type: array
          items:
            type: string
          example: ["taskCreated", "taskUpdated"]

    WebhookList:
      type: object
      properties:
        webhooks:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'

paths:
  /team/{team_id}/webhook:
    post:
      summary: Create a new webhook
      description: Creates a new webhook for the specified team
      tags:
        - webhooks
      parameters:
        - name: team_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '200':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
    get:
      summary: List team webhooks
      description: Retrieves all webhooks for the specified team
      tags:
        - webhooks
      parameters:
        - name: team_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookList'